authorisation

+++ server/routes/users.js
@@ -2,11 +2,11 @@ const router = require("express").Router();
 const bcrypt = require("bcrypt");
 const User = require("../models/User");
 const Game = require("../models/Game")
-
+const protect = require("../middleware/auth")
 
 // update user
-router.put("/:id", async(req, res) => {
-    if(req.body._id == req.params.id || req.body.isAdmin){
+router.put("/:id", protect, async(req, res) => {
+    if(req.user._id == req.params.id || req.user.isAdmin){
         try{
             await User.findByIdAndUpdate(req.params.id, {
                 $set: {
@@ -26,8 +26,8 @@ router.put("/:id", async(req, res) => {
 });
 
 // update user password
-router.put("/updatePassword/:id", async(req,res) => {
-    if(req.body.userId == req.params.id || req.body.isAdmin){
+router.put("/updatePassword/:id", protect, async(req,res) => {
+    if(req.user._id == req.params.id || req.user.isAdmin){
         try{
             const user = await User.findById(req.params.id)
             const validPassword = await bcrypt.compare(req.body.oldPassword, user.password);
@@ -56,8 +56,8 @@ router.put("/updatePassword/:id", async(req,res) => {
 });
 
 // delete user
-router.delete("/:id", async(req, res) => {
-    // if(req.body.userId == req.params.id || req.body.isAdmin){
+router.delete("/:id", protect, async(req, res) => {
+    if(req.user._id == req.params.id){
         try{
             await User.findByIdAndDelete(req.params.id);
             res.status(200).json("Account has been deleted")
@@ -65,9 +65,9 @@ router.delete("/:id", async(req, res) => {
         catch(err){
             return res.status(500).json(err);
         }
-    // } else {
-    //     return res.status(403).json("You can only delete your account")
-    // }
+    } else {
+        return res.status(403).json("You can only delete your account")
+    }
 });
 
 // get user
@@ -87,8 +87,8 @@ router.get("/", async(req, res) => {
 });
 
 // follow user
-router.put("/:id/follow", async(req, res) => {
-    if(req.body.userId !== req.params.id) {
+router.put("/:id/follow", protect, async(req, res) => {
+    if(req.user._id !== req.params.id) {
         try{
             const user = await User.findById(req.params.id);
             const currentUser = await User.findById(req.body.userId);
@@ -109,8 +109,8 @@ router.put("/:id/follow", async(req, res) => {
 });
 
 // unfollow user
-router.put("/:id/unfollow", async(req, res) => {
-    if(req.body.userId !== req.params.id) {
+router.put("/:id/unfollow", protect, async(req, res) => {
+    if(req.user._id !== req.params.id) {
         try{
             const user = await User.findById(req.params.id);
             const currentUser = await User.findById(req.body.userId);
@@ -131,9 +131,9 @@ router.put("/:id/unfollow", async(req, res) => {
 });
 
 // get following
-router.get("/following/:user", async (req, res) => {
+router.get("/following/:user", protect, async (req, res) => {
     try {
-        const currentUser =
+        const currentUser = User.findById(req.user._id)
         await User.findOne({username: req.params.user}) || await User.findById(req.params.user);
         const following = await Promise.all(
             currentUser.following.map((userId) => {
@@ -150,8 +150,7 @@ router.get("/following/:user", async (req, res) => {
 // get followers
 router.get("/followers/:user", async (req, res) => {
     try {
-        const currentUser =
-        await User.findOne({username: req.params.user}) || await User.findById(req.params.user);
+        const currentUser = await User.findOne({username: req.params.user}) || await User.findById(req.params.user);
         const followers = await Promise.all(
             currentUser.followers.map((userId) => {
                 return User.findById(userId)
@@ -167,8 +166,7 @@ router.get("/followers/:user", async (req, res) => {
 // get users favourite games
 router.get("/favourite-games/:user", async (req, res) => {
     try {
-      const currentUser = 
-      await User.findOne({username: req.params.user}) || await User.findById(req.params.user);
+      const currentUser = await User.findOne({username: req.params.user}) || await User.findById(req.params.user);
         const favouriteGames = await Promise.all(
             currentUser.favGames.map((gameId) => {
                 return Game.findById(gameId)
@@ -183,8 +181,7 @@ router.get("/favourite-games/:user", async (req, res) => {
 // get users recently played games
 router.get("/recently-played-games/:user", async (req, res) => {
     try {
-        const currentUser = 
-        await User.findOne({username: req.params.user}) || await User.findById(req.params.user);
+        const currentUser = await User.findOne({username: req.params.user}) || await User.findById(req.params.user);
           const recentlyPlayedGames = await Promise.all(
               currentUser.recentGames.map((gameId) => {
                   return Game.findById(gameId)
@@ -197,9 +194,9 @@ router.get("/recently-played-games/:user", async (req, res) => {
 })
 
 // add game to favourites
-router.put("/:gameId/add-to-favourites", async (req, res) => {
+router.put("/:gameId/add-to-favourites/", protect, async (req, res) => {
     try {
-        const user = await User.findById(req.body.userId)
+        const user = await User.findById(req.user._id)
         if(!user.favGames.includes(req.params.gameId)) {
             await user.updateOne({$push:{favGames:req.params.gameId}})
             res.status(200).json("Game has been added to favourites")
@@ -213,9 +210,9 @@ router.put("/:gameId/add-to-favourites", async (req, res) => {
 })
 
 // remove game from favourites
-router.put("/:gameId/remove-from-favourites", async (req, res) => {
+router.put("/:gameId/remove-from-favourites", protect, async (req, res) => {
     try {
-        const user = await User.findById(req.body.userId)
+        const user = await User.findById(req.user._id)
         if(user.favGames.includes(req.params.gameId)) {
             await user.updateOne({$pull:{favGames:req.params.gameId}})
             res.status(200).json("Game has been removed from favourites")
@@ -230,9 +227,9 @@ router.put("/:gameId/remove-from-favourites", async (req, res) => {
 })
 
 // add game to recently played
-router.put("/:gameId/add-to-recently-played", async (req, res) => {
+router.put("/:gameId/add-to-recently-played", protect, async (req, res) => {
     try {
-        const user = await User.findById(req.body.userId)
+        const user = await User.findById(req.user._id)
         await user.updateOne({$push:
             {recentGames:{
                 $each:[req.params.gameId],
@@ -247,9 +244,9 @@ router.put("/:gameId/add-to-recently-played", async (req, res) => {
 })
 
 // remove game from recently played
-router.put("/:gameId/remove-from-recently-played", async (req, res) => {
+router.put("/:gameId/remove-from-recently-played", protect, async (req, res) => {
     try {
-        const user = await User.findById(req.body.userId)
+        const user = await User.findById(req.user._id)
         if(user.recentGames.includes(req.params.gameId)) {
             await user.updateOne({$pull:{recentGames:req.params.gameId}})
             res.status(200).json("Game removed from recently played")
@@ -261,9 +258,9 @@ router.put("/:gameId/remove-from-recently-played", async (req, res) => {
 })
 
 // add game to last played
-router.put("/:gameId/last-played", async (req, res) => {
+router.put("/:gameId/last-played", protect, async (req, res) => {
     try {
-        const user = await User.findById(req.body.userId)
+        const user = await User.findById(req.user._id)
         await user.updateOne({$set:{lastPlayed:req.params.gameId}})
         res.status(200).json("Game set to last played")
     }
@@ -273,9 +270,9 @@ router.put("/:gameId/last-played", async (req, res) => {
 })
 
 // update points
-router.put("/add-user-points/:points", async (req, res) => {
+router.put("/add-user-points/:points", protect, async (req, res) => {
     try {
-        const user = await User.findById(req.body.userId)
+        const user = await User.findById(req.user._id)
         await user.updateOne({$inc:{points:req.params.points}})
         res.status(200).json("Points updated")
     }
@@ -285,9 +282,9 @@ router.put("/add-user-points/:points", async (req, res) => {
 })
 
 // add top score
-router.put("/add-top-score/:score", async (req, res) => {
+router.put("/add-top-score/:score", protect, async (req, res) => {
     try {
-        const user = await User.findById(req.body.userId)
+        const user = await User.findById(req.user._id)
         const checkGame = user.topScores.find(score => score.gameId === req.body.gameId)
         if(checkGame) return res.json("game already added")
         await user.updateOne({$push:{
@@ -303,10 +300,10 @@ router.put("/add-top-score/:score", async (req, res) => {
 })
 
 // update top score
-router.put("/update-top-score/:score", async(req, res) => {
+router.put("/update-top-score/:score", protect, async(req, res) => {
     try {
         await User.findOneAndUpdate({
-            _id: req.body.userId,
+            _id: req.user._id,
             'topScores.gameId': req.body.gameId
         },
         {$set: {'topScores.$.score': req.params.score}},
